TODO: ImageJ Launcher: support finding jars/ and plugins/ outside .app/
Our current application .zip has only stubs for MacOSX. It seems to be
the intention to leave the jars/ directory outside of the .app.

This has several advantages over the approach Fiji took (everything inside
the .app/ directory):

1) it is easy to put new stuff into the jars/ and plugins/ directories,

2) it is obvious how to have 32-bit and 64-bit MacOSX versions without
   duplicating the .jar files

3) users can inspect the components of ImageJ more conveniently.

However, there is also a drawback: it is no longer easy to quickly install
ImageJ into the Applications/ folder.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>